tpm: Don't propagate TPM measurement errors to the verifiers layer
Currently if the EFI firmware fails to do a TPM measurement for a file,
the error will be propagated to the verifiers framework and so opening
the file will not succeed.

This mean that buggy firmwares will prevent the system to boot since the
loader won't be able to open any file. But failing to do TPM measurements
shouldn't be a fatal error and the system should still be able to boot.

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>